# Conflicts:
#	.travis.yml
#	Docs/AngelScriptAPI.h
#	Docs/LuaScriptAPI.dox
#	Docs/ScriptAPI.dox
#	README.md
#	Source/Urho3D/AngelScript/APITemplates.h
#	Source/Urho3D/AngelScript/CoreAPI.cpp
#	Source/Urho3D/AngelScript/GraphicsAPI.cpp
#	Source/Urho3D/AngelScript/IOAPI.cpp
#	Source/Urho3D/AngelScript/InputAPI.cpp
#	Source/Urho3D/AngelScript/MathAPI.cpp
#	Source/Urho3D/AngelScript/ResourceAPI.cpp
#	Source/Urho3D/AngelScript/UIAPI.cpp
#	Source/Urho3D/CMakeLists.txt
#	Source/Urho3D/Core/Object.cpp
#	Source/Urho3D/Core/Object.h
#	Source/Urho3D/Graphics/DebugRenderer.cpp
#	Source/Urho3D/Graphics/DebugRenderer.h
#	Source/Urho3D/Graphics/RenderPath.cpp
#	Source/Urho3D/IO/FileSystem.cpp
#	Source/Urho3D/IO/FileSystem.h
#	Source/Urho3D/LuaScript/pkgs/Core/Object.pkg
#	Source/Urho3D/LuaScript/pkgs/Core/Timer.pkg
#	Source/Urho3D/LuaScript/pkgs/Graphics/DebugRenderer.pkg
#	Source/Urho3D/LuaScript/pkgs/Graphics/Graphics.pkg
#	Source/Urho3D/LuaScript/pkgs/Graphics/RenderPath.pkg
#	Source/Urho3D/LuaScript/pkgs/IO/FileSystem.pkg
#	Source/Urho3D/LuaScript/pkgs/IO/Log.pkg
#	Source/Urho3D/LuaScript/pkgs/Input/InputEvents.pkg
#	Source/Urho3D/LuaScript/pkgs/Math/Color.pkg
#	Source/Urho3D/LuaScript/pkgs/Math/StringHash.pkg
#	Source/Urho3D/LuaScript/pkgs/Resource/Image.pkg
#	Source/Urho3D/LuaScript/pkgs/Resource/XMLElement.pkg
#	Source/Urho3D/LuaScript/pkgs/UI/ScrollView.pkg
#	Source/Urho3D/LuaScript/pkgs/UI/UIElement.pkg
#	Source/Urho3D/Math/Rect.h
#	Source/Urho3D/Math/Vector2.h
#	Source/Urho3D/Resource/Image.cpp
#	Source/Urho3D/Resource/XMLElement.cpp
#	Source/Urho3D/Resource/XMLElement.h
#	Source/Urho3D/Scene/Serializable.cpp
#	Source/Urho3D/Scene/Serializable.h

+++ Source/Urho3D/Graphics/View.cpp
@@ -2531,7 +2531,7 @@ void View::SetupShadowCameras(LightQueryResult& query)
 {
     Light* light = query.light_;
 
-    int splits = 0;
+    unsigned splits = 0;
 
     switch (light->GetLightType())
     {
@@ -2605,7 +2605,7 @@ void View::SetupShadowCameras(LightQueryResult& query)
         break;
     }
 
-    query.numSplits_ = (unsigned)splits;
+    query.numSplits_ = splits;
 }
 
 void View::SetupDirLightShadowCamera(Camera* shadowCamera, Light* light, float nearSplit, float farSplit)
